// WalletAuth/src/sdk/walletInfraProvider/interfaces.d.ts
export type CoreKitMode = UX_MODE_TYPE | "nodejs" | "react-native";
export type V3TSSLibType = {
    keyType: string;
    lib: unknown;
};
export type V4TSSLibType = typeof TssFrostLib | typeof TssDklsLib;
export type TssLibType = V4TSSLibType | V3TSSLibType;
export interface IStorage {
    getItem(key: string): string | null;
    setItem(key: string, value: string): void;
}
export interface IAsyncStorage {
    async?: boolean;
    getItem(key: string): Promise<string | null>;
    setItem(key: string, value: string): Promise<void>;
}
export type SupportedStorageType = "local" | "session" | "memory" | IStorage;
export interface InitParams {
    /**
     * @defaultValue `true`
     * handle the redirect result during init()
     */
    handleRedirectResult: boolean;
    /**
     * @defaultValue `true`
     * rehydrate the session during init()
     */
    rehydrate?: boolean;
}
export interface BaseLoginParams {
    serverTimeOffset?: number;
}
export interface SubVerifierDetailsParams extends BaseLoginParams {
    subVerifierDetails: SubVerifierDetails;
}
export interface AggregateVerifierLoginParams extends BaseLoginParams {
    aggregateVerifierIdentifier: string;
    subVerifierDetailsArray: SubVerifierDetails[];
    aggregateVerifierType?: AGGREGATE_VERIFIER_TYPE;
}
// ...
export type OAuthLoginParams = (SubVerifierDetailsParams | AggregateVerifierLoginParams) & {
    importTssKey?: string;
};
// ...
export interface JWTLoginParams {
    /**
     * Name of the verifier created on Web3Auth Dashboard. In case of Aggregate Verifier, the name of the top level aggregate verifier.
     */
    verifier: string;
    /**
     * Unique Identifier for the User. The verifier identifier field set for the verifier/ sub verifier. E.g. "sub" field in your on jwt id token.
     */
    verifierId: string;
    /**
     * The idToken received from the Auth Provider.
     */
    idToken: string;
    /**
     * Name of the sub verifier in case of aggregate verifier setup. This field should only be provided in case of an aggregate verifier.
     */
    subVerifier?: string;
    /**
     * Extra verifier params in case of a WebAuthn verifier type.
     */
    extraVerifierParams?: PasskeyExtraParams;
    /**
     * Any additional parameter (key value pair) you'd like to pass to the login function.
     */
    additionalParams?: ExtraParams;
    /**
     * Key to import key into Tss during first time login.
     */
    importTssKey?: string;
    /**
     * Number of TSS public keys to prefetch. For the best performance, set it to
     * the number of factors you want to create. Set it to 0 for an existing user.
     * Default is 1, maximum is 3.
     */
    prefetchTssPublicKeys?: number;
}
// ...
export type WEB3AUTH_NETWORK_TYPE = (typeof WEB3AUTH_NETWORK)[keyof typeof WEB3AUTH_NETWORK];
export type USER_PATH_TYPE = (typeof USER_PATH)[keyof typeof USER_PATH];
export interface WalletAuthOptions {
    /**
     * The Web3Auth Client ID for your application. Find one at https://dashboard.web3auth.io
     */
    web3AuthClientId: string;
    /**
     * The threshold signing library to use.
     */
    tssLib: TssLibType;
    /**
     * @defaultValue `false`
     */
    manualSync?: boolean;
    /**
     * @defaultValue `${window.location.origin}/serviceworker`
     */
    baseUrl?: string;
    /**
     *
     * @defaultValue `'sapphire_mainnet'`
     */
    web3AuthNetwork?: WEB3AUTH_NETWORK_TYPE;
    /**
     *  storage for mpc-core-kit's local state.
     *  storage replaces previous' storageKey and asyncStorage options.
     *
     *  Migration from storageKey and asyncStorage to storage guide.
     *
     *  For StorageKey, please replace
     *  - undefined with localStorage
     *  - "local" with localStorage
     *  - "session" with sessionStorage
     *  - "memory" with new MemoryStorage()
     *
     *  For asyncStorage, provide instance of IAsyncStorage.
     *
     */
    storage: IAsyncStorage | IStorage;
    /**
     * @defaultValue 86400
     */
    sessionTime?: number;
    /**
     * @defaultValue `'POPUP'`
     */
    uxMode?: CoreKitMode;
    /**
     * @defaultValue `false`
     * enables logging of the internal packages.
     */
    enableLogging?: boolean;
    /**
     * This option is used to specify the url path where user will be
     * redirected after login. Redirect Uri for OAuth is baseUrl/redirectPathName.
     *
     *
     * @defaultValue `"redirect"`
     *
     * @remarks
     * At verifier's interface (where you obtain client id), please use baseUrl/redirectPathName
     * as the redirect_uri
     *
     * Torus Direct SDK installs a service worker relative to baseUrl to capture
     * the auth redirect at `redirectPathName` path.
     *
     * For ex: While using serviceworker if `baseUrl` is "http://localhost:3000/serviceworker" and
     * `redirectPathName` is 'redirect' (which is default)
     * then user will be redirected to http://localhost:3000/serviceworker/redirect page after login
     * where service worker will capture the results and send it back to original window where login
     * was initiated.
     *
     * For browsers where service workers are not supported or if you wish to not use
     * service workers,create and serve redirect page (i.e redirect.html file which is
     * available in serviceworker folder of this package)
     *
     * If you are using redirect uxMode, you can get the results directly on your `redirectPathName`
     * path using `getRedirectResult` function.
     *
     * For ex: if baseUrl is "http://localhost:3000" and `redirectPathName` is 'auth'
     * then user will be redirected to http://localhost:3000/auth page after login
     * where you can get login result by calling `getRedirectResult` on redirected page mount.
     *
     * Please refer to examples https://github.com/torusresearch/customauth/tree/master/examples
     * for more understanding.
     *
     */
    redirectPathName?: string;
    /**
     * @defaultValue `false`
     * Disables the cloud factor key, enabling the one key semi custodial flow.
     * Recommended for Non Custodial Flow.
     */
    disableHashedFactorKey?: boolean;
    /**
     * @defaultValue `Web3AuthOptions.web3AuthClientId`
     * Overwrites the default value ( clientId ) used as nonce for hashing the hash factor key.
     *
     * If you want to aggregate the mfa status of client id 1 and client id 2  apps
     * set hashedFactorNonce to some common clientID, which can be either client id 1 or client id 2 or any other unique string
     * #PR 72
     * Do not use this unless you know what you are doing.
     */
    hashedFactorNonce?: string;
    serverTimeOffset?: number;
    /**
     * Set this flag to false to generate keys on client side
     * by default keys are generated on using dkg protocol on a distributed network
     * Note: This option is not supported for ed25519 key type
     * @defaultValue `true`
     */
    useDKG?: boolean;
    /**
     * @defaultValue `false` for secp256k1 and `true` for ed25519
     * Set this flag to true to use the client generated key for signing
     * Note: This option is set to true for ed25519 key type by default to ensure ed25519 mpc key  seed exportablity.
     * The seed thn can be used for importing user's key other wallets like phantom etc
     * If you set this flag to false for ed25519 key type, you will not be able to export the seed for ed25519 keys and
     * only scalar will be exported, scalar can be used for signing outside of this sdk but not for importing the key in other wallets.
     */
    useClientGeneratedTSSKey?: boolean;
}